Add `condition` to `DataField` model
While conditions are not currently supported by root data fields, they are supported by complex member field.
As complex member field reuse the `DataField` model, they thus require condition to be supported by `DataField`.
It is conceivable that support for condition is added in the future for root field as well so it makes sense having optional support on `DataField`