Refactor `QuickcaseAuthentication`
Instead of an abstract class with logic-driven implementations, turn `QuickcaseAuthentication` into a value-based class with final fields set via constructor.
The intent is to reduce the differences and unify behaviour as much as possible between user and client-based authentication.

At the same time, the structure of the authentication is also reviewed to better align with Spring Security's philosophy where the authentication class is exclusively intended
for Spring while the contained `UserDetails` (here `UserInfo`) is intended for the consuming applications.
Hence, all user-centric properties are now exclusively captured in `UserInfo` and `QuickcaseAuthentication` extends them with client-centric properties for the purpose of Spring Security.